Ability to save and open PNG files
