+++ src/by/painter/controller/drawer/PaintbrushDrawer.java
@@ -1,10 +1,8 @@
-package by.painter.controller;
+package by.painter.controller.drawer;
 
 import by.painter.view.Viewable;
 
 import java.awt.*;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
 
 public class PaintbrushDrawer extends PenDrawer {
 
@@ -15,7 +13,7 @@ public class PaintbrushDrawer extends PenDrawer {
     @Override
     public void drawFigure(Graphics g) {
         Graphics2D g2d = (Graphics2D) g;
-        g2d.setStroke(new BasicStroke(10.0f, 1, 1));
+        g2d.setStroke(new BasicStroke(10.0f, BasicStroke.CAP_ROUND, BasicStroke.JOIN_ROUND));
         g2d.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
         g2d.setColor(painter.getInstrumentColor());
         g2d.drawLine(x1, y1, x2, y2);
