+++ src/by/painter/controller/drawer/RectangleDrawer.java
@@ -1,34 +1,36 @@
-package by.painter.controller;
+package by.painter.controller.drawer;
 
+import by.painter.controller.DrawingInstrument;
 import by.painter.view.TemporalCanvas;
 import by.painter.view.Viewable;
-
 import java.awt.*;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 
-public class TriangleDrawer extends DrawingInstrument {
+public class RectangleDrawer extends DrawingInstrument {
 
-    protected int x1, y1, x2;
+    protected int x1, y1, x2, y2;
     protected final TemporalCanvas temporalCanvas;
     protected Graphics g;
 
-    public TriangleDrawer(Viewable w) {
+    public RectangleDrawer(Viewable w) {
         super.mainCanvas = w.getMainCanvas();
-        super.adapter = new TriangleAdapter();
+        super.adapter = new RectangleAdapter();
         super.painter = w.getPainter();
         temporalCanvas = new TemporalCanvas(this);
     }
 
     @Override
     public void drawFigure(Graphics g) {
-        int px = Math.abs((x1 + x2) / 2);
-        int py = y1 - (x2 - x1);
+        int px = Math.min(x1, x2);
+        int py = Math.min(y1, y2);
+        int pw = Math.abs(x1 - x2);
+        int ph = Math.abs(y1 - y2);
         g.setColor(painter.getInstrumentColor());
-        g.drawPolygon(new int[]{x1, x2, px}, new int[]{y1, y1, py}, 3);
+        g.drawRect(px, py, pw, ph);
     }
 
-    private class TriangleAdapter extends MouseAdapter {
+    private class RectangleAdapter extends MouseAdapter {
 
         @Override
         public void mousePressed(MouseEvent e) {
@@ -43,6 +45,7 @@ public class TriangleDrawer extends DrawingInstrument {
         @Override
         public void mouseDragged(MouseEvent e) {
             x2 = e.getX();
+            y2 = e.getY();
             temporalCanvas.repaint();
         }
 
@@ -50,6 +53,7 @@ public class TriangleDrawer extends DrawingInstrument {
         public void mouseReleased(MouseEvent e) {
             mainCanvas.remove(temporalCanvas);
             x2 = e.getX();
+            y2 = e.getY();
             drawFigure(g);
             mainCanvas.repaint();
         }
